app/src/main/java/ru/olegivo/afs/common/domain/DateTimeUtilsImpl.kt
/*
 * Copyright (C) 2020 Oleg Ivashchenko <olegivo@gmail.com>
 *
 * This file is part of AFS.
 *
 * AFS is free software: you can redistribute it and/or modify
 * it under the terms of the MIT License.
 *
 * AFS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * AFS.
 */

package ru.olegivo.afs.common.domain

import org.threeten.bp.DateTimeUtils
import org.threeten.bp.OffsetDateTime
import org.threeten.bp.ZoneId
import org.threeten.bp.temporal.WeekFields
import java.util.Date
import javax.inject.Inject

class DateTimeUtilsImpl @Inject constructor() :
    ru.olegivo.afs.common.domain.DateTimeUtils {

    override fun getWeekDayNumber(date: Date): Int {
        val instant = DateTimeUtils.toInstant(date)
        val offsetDateTime = OffsetDateTime.ofInstant(instant, ZoneId.systemDefault())
        return offsetDateTime[WeekFields.ISO.dayOfWeek()]
    }
}
